遥控改_依据速度动态KpKi

+++ HARDWARE/CDHData/cdhdata.h
@@ -52,6 +52,6 @@ Fencha_struct FindSpace1(u8 dst,u8 num,u8 dir);//8位处理检测到磁导航的数据函数
 void PID_AUTO_QianLun(u16 j_speed,float kp,float ki,float kd);
 void PID_AUTO_HouLun(u16 j_speed2,float kp2,float ki2,float kd2);	//后轮的PID调节
 
-void PID_SD_Adjust(u16 j_speed,float kp,float ki,float kd);
+void PID_SD_Adjust(u16 j_speed,float kp,u16 ki,float kd);
 
 #endif
